requirements.txt
Click==7.0
cycler==0.10.0
Flask==1.1.1
Flask-Uploads==0.2.1
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.18.0
opencv-python==4.1.2.30
pandas==0.25.3
Pillow==6.2.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scipy==1.4.1
seaborn==0.9.0
six==1.13.0
Werkzeug==0.16.0